:bug: 순환참조 일어나는 곳 수정
